+++ backend/src/main/java/com/application/quai/model/entity/Menu.java
@@ -27,7 +27,7 @@ public class Menu{
 
  @Id
  @GeneratedValue(strategy = GenerationType.IDENTITY)
- private int idMenu;
+ private Long idMenu;
 
  @NotNull
  @Column(nullable=false)
